+++ src/main/java/pl/smsapi/api/action/contacts/groups/GroupEdit.java
@@ -1,0 +1,43 @@
+package pl.smsapi.api.action.contacts.groups;
+
+import org.json.JSONObject;
+import pl.smsapi.api.action.AbstractAction;
+
+public class GroupEdit extends AbstractAction<Group> {
+
+    private final String id;
+
+    public GroupEdit(String id) {
+        this.id = id;
+    }
+
+    @Override
+    protected String endPoint() {
+        return "contacts/groups/" + id;
+    }
+
+    @Override
+    protected String httpMethod() {
+        return "PUT";
+    }
+
+    public GroupEdit setName(String name) {
+        params.put("name", name);
+        return this;
+    }
+
+    public GroupEdit setDescription(String description) {
+        params.put("description", description);
+        return this;
+    }
+
+    public GroupEdit setIdx(String idx) {
+        params.put("idx", idx);
+        return this;
+    }
+
+    @Override
+    protected Group createResponse(String data) {
+        return new Group.GroupFromJsonFactory().createFrom(new JSONObject(data));
+    }
+}