+++ src/main/java/pl/smsapi/api/response/contacts/ContactsPermissionListResponse.java
@@ -4,6 +4,10 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import pl.smsapi.api.response.ListResponse;
 
+/**
+ * @deprecated use {@link pl.smsapi.api.action.contacts.groups.Permissions()} instead
+ */
+@Deprecated
 public class ContactsPermissionListResponse extends ListResponse<ContactsPermissionResponse> {
 
     public ContactsPermissionListResponse(int count, JSONArray jsonArray) {